fix: includes logo
