+++ frontend/scripts/react-components/shared/nav/nav.container.js
@@ -1,10 +1,13 @@
 import { connect } from 'react-redux';
+import routerLinks from 'router/nav-links';
 import Nav from './nav.component';
 
 function mapStateToProps(state) {
   const { type, routesMap } = state.location;
   return {
-    ...routesMap[type].nav
+    ...routesMap[type].nav,
+    links: routerLinks.nav,
+    showLogo: type !== 'home'
   };
 }
