+++ frontend/scripts/react-components/static-content/static-content.container.js
@@ -1,13 +1,13 @@
 import { connect } from 'react-redux';
 import { NOT_FOUND } from 'redux-first-router';
+import routerLinks from 'router/nav-links';
 import StaticContent from './static-content.component';
-import sidebarLinks from './nav-sidebar-links';
 import { getStaticContentFilename } from './static-content.actions';
 
 function mapStateToProps(state) {
   const { staticContent, location } = state;
   const filename = getStaticContentFilename(location);
-  const links = sidebarLinks[filename] || sidebarLinks.default;
+  const links = routerLinks[filename] || routerLinks.navSidebar;
   return {
     content: staticContent.markdown[filename],
     links,
